+++ src/core/plugins/core-death/shared/config.ts
@@ -1,3 +1,0 @@
-export const DEATH_CONFIG = {
-    RESPAWN_KEY: 69, // E
-};