+++ src/core/plugins/core-death/shared/events.ts
@@ -1,4 +1,0 @@
-export const DEATH_EVENTS = {
-    UPDATE_DEATH_TIMER_MS: 'update:death:time:left',
-    RESPAWN_PRESSED: 'death:event:server:respawn:pressed',
-};