+++ src/core/server/api/consts/constDefaultSystems.ts
@@ -1,0 +1,5 @@
+import { DefaultDeathSystem } from '@AthenaServer/systems/defaults/death';
+
+export const defaultSystemsConst = {
+    death: DefaultDeathSystem,
+};