+++ src/core/server/api/consts/constSystem.ts
@@ -7,9 +7,11 @@ import { PluginSystem } from '../../systems/plugins';
 import { SoundSystem } from '../../systems/sound';
 import { StorageSystem } from '../../systems/storage';
 import { World } from '../../systems/world';
+import { defaultSystemsConst } from './constDefaultSystems';
 
 export const systemConst = {
     agenda: AgendaSystem,
+    default: defaultSystemsConst,
     itemFactory: ItemFactory,
     effects: ItemEffects,
     identifier: Identifier,